move ConfigurationLoader to config package for consistency
